+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-03-13
+
+### Added
+
+- example using JWT authentication
+
+### Changed
+
+- NitramInner is not required as a resource
+- `WSSessionAnonymResource::auth()` no longer takes a `session_id` parameter — the WebSocket session ID is now used internally as the session ID
+- `UserSession::id` type changed from `String` to `Uuid`
+
+### Changed
+
+- websocket server URL is now passed to the constructor instead of being read from environment variables. VITE_WS_SERVER is no longer required.
+
 ## [0.3.5] - 2025-12-14
 
 ### Added